requirements.txt
Django==5.0.6
django-colorfield==0.11.0
django-admin-interface==0.28.8
djangorestframework==3.15.2
django-rest-swagger==2.2.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9